Drop use of 2to3
Rewrote the source to be compatible with all Python versions (with the
help of python-modernize, and, when that proved to be woefully
inadequate, 2to3).

Introduced a dependency on six, because I can't be bothered to figure
out how to implement six.reraise() on my own.

+++ src/zope/testrunner/shuffle.py
@@ -31,7 +31,7 @@ class Shuffle(zope.testrunner.feature.Feature):
             # We can't rely on the random modules seed initialization because
             # we can't introspect the seed later for reporting.  This is a
             # simple emulation of what random.Random.seed does anyway.
-            self.seed = long(time.time() * 256) # use fractional seconds
+            self.seed = int(time.time() * 256) # use fractional seconds
 
     def global_setup(self):
         rng = random.Random(self.seed)
